Add RPC API account balancemulti
Why:

* After a first iteration of the account balance  API endpoint was
merged into master, we noticed Etherscan has a separate action for
getting the Ether balance of multiple addresses in a single call,
`action=balancemulti`. This commit adds support for this action.
* Issue link: https://github.com/poanetwork/poa-explorer/issues/138

This change addresses the need by:

* Editing `API.RPC.AddressControllerTest` to have coverage for
`action=balancemulti` calls and make sure `action=balance` works as we'd
like when multiple addresses are given.
* Adding `balancemulti/2` to `API.RPC.AddressController`.