Catch up index based on whether previous catch up found missing blocks
When a catchup_index task completes without error, increase the interval
until the next check by 2 if no blocks were missing and decrease it by 2
if blocks were missing.  Minimum interval is the same as the realtime
interval while the maximum interval is arbitarily chosen to be 10x the
realtime interval.