Use Path.absname so that log paths are stable during reload
Fixes #628

The Phoenix.CodeReloader causes the file loggers to restart and reload
their config, but when they reload, the current working directory has
shifted from the project root to the OTP app root, so it produces log
files under each OTP app (`apps/APP_NAME/logs`).  We don't want this, so
compute the absolute path in the config files, so that the loaded
`Application` environment has an absolute path in it that can't shift
when the working directory changes.