Temporary work-around for #673 value-only transactions
The gas_used == gas heuristic means failures pre-Byzantium (and on ETC)
is only a heuristic: it fails for value-only transactions where ETC
submitters are using exactly the transaction base fee (21,000 gas)
often.  Because this is so common, we'll use this work-around now until
we can restructure EthereumJSONRPC and Indexer to verify pre-Byzantium
status using Internal Transactions.