Fix token balances' query
This query was considering the next-to-last token balance when the last
one had the value 0.

In order to this work properly, I'm moving the query that gets token
balances from the last block to a sub-query and moving the condition to
consider just values greater than 0 to the main query.  This way we are
sure that we always are considering the last block and we are skipping
the ones that have the value equal to 0.