Don't use closure for BufferedTask Task.Supervisor.async_nolink call
Fixes #676

For some reason, which I haven't tracked down yet, a closure inside of
`BufferedTask.spawn_next_batch/1` continually grows in size.  When this
closure is passed to `Task.Supervisor.async_nolink`, the
`Task.Supervisor`, for #676, this was `Indexer.TaskSupervisor` and is
now `Indexer.InternalTransaction.TaskSupervisor`, it also grows in
memory size.  I think this is related to the way that functions and
closures are serialized when sent between processes, but I haven't
figured out how to verify that.

The change also cleared the MsgQ for `Indexer.TransactionFetcher`,
probably because its not blocked by the call to
`Indexer.InternalTransaction.Fetcher`, which was blocked due to
something to do with the memory usage of the closures.

The `Indexer.InternalTransaction.TaskSupervisor` consistently shows
multiple tasks under it in `:observer`, which it did not previously
either.