Fix token holders' tab at the Token's page
This page was breaking when the given Token has the total supply nil.

To fix that, we added a view function to check if we should show the
total supply percentage. Since we want to display just for Tokens that
have total supply, we are ignoring for the ones that don't have.