Shrinkable BufferedTasks
`Indexer.BufferedTask` now uses an `Indexer.BoundQueue` instead of a
`:queue`.  All `Indexer.BufferedTask`s register with the memory monitor
passed in to `BufferedTask.init/1` and will shrink their bound queue
when asked.  If they run out of work and have shrunk `c:init/2` will be
called again on the callback module.

`Indexer.Logger.process/1` will generate iodata of the pid and
registered name of a process if available.