Add nonce field to Address object and addresses table.
This caches the highest nonce value seen for sent transactions for a given address.

The top addresses query no longer needs to scan the transactions table for nonce values.

The database migration script adds the nonce field to the addresses table and sets
the nonce value according to the contents of the transactions table.