Only query for coin balance when channel connected
Normally, it is good idea to only query the database once no matter the
number of sockets for a channel, but most addresses don't have listeners,
so it is cheaper to not query unless the channel is in use.

If we start to see popular addresses, con_cache can be used to only do
the query once across all connected sockets for the same address.

Co-authored-by: Chris McCord <chris@chrismccord.com>