Enforce DB transaction's order to prevent deadlocks
Problem: due to the ShareLock mechanism used by PostgreSQL when there are several DB transactions are acting on multiple rows of the same table, it's possible to incur in a deadlock and so into an error.

Solution: enforce a consistent order of ShareLock acquisition in all the relevant transaction.
Also add clear documentation that should help in the future.